feat: spring security

+++ src/main/java/com/task/taskmanager/infrastructure/repository/StateRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.task.taskmanager.domain.entity.State;
 
-public interface StateRepository extends JpaRepository<State, Integer>{
+public interface StateRepository extends JpaRepository<State, Long>{
 
 }
